requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
google-generativeai
